requirements.txt
opencv-python
numpy
Pillow
websockets
mss
scikit-learn